Migrate: change filter.event_type from str to list
