+++ src/argus/notificationprofile/serializers.py
@@ -1,8 +1,33 @@
 from rest_framework import fields, serializers
 
-from .primitive_serializers import FilterBlobSerializer, FilterPreviewSerializer
+from argus.incident.constants import INCIDENT_LEVELS
+from argus.incident.models import Event
+from .primitive_serializers import CustomMultipleChoiceField, FilterPreviewSerializer
 from .media import api_safely_get_medium_object
 from .models import DestinationConfig, Filter, Media, NotificationProfile, TimeRecurrence, Timeslot
+
+
+class FilterBlobSerializer(serializers.Serializer):
+    sourceSystemIds = serializers.ListField(
+        child=serializers.IntegerField(min_value=1),
+        allow_empty=True,
+        required=False,
+    )
+    tags = serializers.ListField(
+        child=serializers.CharField(min_length=3),
+        allow_empty=True,
+        required=False,
+    )
+    open = serializers.BooleanField(required=False, allow_null=True)
+    acked = serializers.BooleanField(required=False, allow_null=True)
+    stateful = serializers.BooleanField(required=False, allow_null=True)
+    maxlevel = serializers.IntegerField(
+        required=False, allow_null=True, max_value=max(INCIDENT_LEVELS), min_value=min(INCIDENT_LEVELS)
+    )
+    event_types = CustomMultipleChoiceField(
+        choices=Event.Type.choices,
+        required=False,
+    )
 
 
 class TimeRecurrenceSerializer(serializers.ModelSerializer):
